[font-types] Get rid of OffsetLen and OffsetHot
OffsetLen can live in compile-tables, maybe, and OffsetHost no longer
makes sense.
