+++ font-tables/src/lib.rs
@@ -14,8 +14,7 @@ pub mod parse_prelude {
 
     pub use font_types::{
         BigEndian, F2Dot14, FWord, Fixed, FixedSized, LongDateTime, MajorMinor, Offset, Offset16,
-        Offset24, Offset32, OffsetHost, OffsetLen, ReadScalar, Scalar, Tag, UfWord, Uint24,
-        Version16Dot16,
+        Offset24, Offset32, ReadScalar, Scalar, Tag, UfWord, Uint24, Version16Dot16,
     };
     pub use std::ops::Range;
 }
